bsc: hack around 2-port BRAM
We need a BRAM with one RW and one RO port. Using BRAM2(Load) for this
as-is doesn't work even if one port is only ever read - Yosys can't
infer that. assumes the memory is true dual-port and can't techmap it
onto anything.

This disables using BRAM2*.v from the Bluespec stdlib and pulls in a
modified copy that yanks out the write enable on port 0.